Replaced earlier emacs configuration file with current file
This commit also introduces Emacs as a parent for the dotemacs
config. This will help down the line when the website is used to share
more Emacs related notes that do not fit on a blog post.

The miscreant export file blog.md has also been deleted. THis file was a
mistaken export containing the entire blog file.