Fixes tags and publish date of last post
ox-hugo automatically sets the publish date using the special closed
property of a TODO. This is convenient and preferable. For some reason,
the option was not enabled in my updated Emacs configuration and that is
rectified now. Now posts are listed by their published date, and the
last modified date is available as meta data for reference.